pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=75.8.2"]

[project]
authors = [
  {name = "Kairos Research", email = "labs@kairosresearch.io"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "eth-account>=0.13.7",
  "eth-utils>=5.3.0",
  "httpx>=0.28.1",
  "ipython>=9.3.0",
  "msgpack>=1.1.0",
  "nest-asyncio>=1.6.0",
  "polars>=1.27.1",
  "requests>=2.32.3",
  "websockets>=15.0.1",
]
description = "A lightweight Python package providing minimal trading functions for selected exchanges, organized by exchange-specific modules."
license = {text = "MIT"}
name = "krex"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"

[project.urls]
Repository = "https://github.com/kairos-research-ltd/krex"

[dependency-groups]
dev = [
  "ruff>=0.9.9",
  "pre-commit>=4.1.0",
  "asyncio>=3.4.3",
  "python-dotenv>=1.1.0",
  "pytest-asyncio>=0.26.0",
]

[[tool.uv.index]]
explicit = false
name = "kairosresearch"
publish-url = "https://kairosresearch.io/pypi/"
url = "https://kairosresearch.io/pypi/simple/"
